fix: constants again

+++ include/constants.hpp
@@ -37,7 +37,7 @@ const int ALTURA_BASE = 50;                  // Altura da base (pixels)
 constexpr int NUM_CANOS = 4;                 // Numero de canos renderizados simultaneamente
 constexpr int VELOCIDADE_CANO = 3;           // Velocidade do Cano (pixels/frame)
 constexpr int ESPACO_ENTRE_OS_CANOS = 250;   // Espaço entre os Canos (pixels)
-const int GRAVIDADE = 800.0f;                // Define a aceleração da gravidade
-const int JUMPSPEED = -300.0f;               // Define a velocidade instantanea ao pular
+const float GRAVIDADE = 800.0f;              // Define a aceleração da gravidade
+const float JUMPSPEED = -300.0f;             // Define a velocidade instantanea ao pular
 
 #endif
